:nut_and_bolt: Updating to beta 0.1.073
The main change that would prevent this from building is the fact that constant declarations are not allowed to be forward referenced inside procedure bodies. Which for us mostly means that you can't put local procedures inside procedures at the bottom (which I tended to do quite often). These have been moved to be declared before they are used.

Also changing the pointer dereferences from using `<<` to use `.*` since the old way has been hinted to be phased out eventually.

Both of these changes actually happened in beta 0.1.065, but I've verified that we're building and running in 0.1.073 too.