:thread: :parachute: More options for string<->float conversions
Now the parser can take floating point numbers directly in their float64 binary representation as in `#0x3ff0000000000000` without having the parsable floating point number preceding it as in `1.0 #0x3ff0000000000000`.

The unparser now takes a setting to output either only the floating point number, only the f64 binary representation or both.

For large arrays of floating point numbers that aren't meant to be modified by hand on the file, you might want to just to the binary option only, since it's more precise when parsing and unparsing as well as faster.