:kite: *Breaking* float parsing changes!
With this change, we now either pass floating point numbers either as a literal float or as a hexadecimal number that has the binary representation of it.

Before we used to support having both, and then we had to try to disambiguate if the float and hex representations differed from each other to tell the user, in case they changed the floating point number but left the hex stuff. That was a mess, cause it's not trivial to tell if the user ligitimately did a change or if it was due to imprecission when outputting/parsing the float literal.

Instead, if you now want to use the binary representation you can spit it with a comment that tells you the actual float value, but the binary will just get used in that case. This makes it much more clear that if you wanna modify it, you have to remove the binary representation.

This also makes float parsing if using binary way faster, because we just read a u64 literal in hexadecimal and avoid the slow float parsing altogether.

This change can make it so you get an error if you try to parse the old format that had the float followed by the binary, since we now only take one of those.