:dependabot: Splitting tokenizer and parser
This module now depends on [osor_tokenizer](https://github.com/osor-io/osor_tokenizer). So you'll need to make that available as well to use it. :broken_heart:

There was enough other modules that only required the standalone tokenizer that it makes sense for those to be separate.

I'd like for it to be a way to bundle up these modules together, but this is getting into dependency-management and that's a big yikes. At the end of the day, it's just code, so if you can get this module you can always get the other one.